+++ app.js
@@ -8,6 +8,7 @@ const session = require('express-session');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
 const indexRouter = require('./routes/index');
 const usersRouter = require('./routes/users');
+const listRouter = require('./routes/lists')
 const { sessionSecret } = require('./config');
 const { restoreUser } = require('./auth');
 
@@ -47,6 +48,7 @@ store.sync();
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
+app.use('/lists', listRouter)
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
