+++ auth.js
@@ -17,7 +17,7 @@ const requireAuth = (req, res, next) => {
 const restoreUser = async (req, res, next) => {
     // Log the session object to the console
     // to assist with debugging.
-    console.log(req.session);
+    //console.log(req.session);
 
     if (req.session.auth) {
         const { userId } = req.session.auth;
